+++ src/Metadata/JobMetadata.cs
@@ -17,7 +17,7 @@ namespace Hangfire.Dashboard.Management.v2.Metadata
 		public MethodInfo MethodInfo { get; set; }
 
 		public string MethodName => Type.Name + "_" + MethodInfo.Name;
-		public string JobId => $"{MenuName}/{Name.ScrubURL()}";
+		public string JobId => $"{MenuName}/{MethodName.ScrubURL()}";
 		public string Name => $"{DisplayName ?? MethodName}";
 	}
 }