+++ src/Support/JobsHelper.cs
@@ -14,6 +14,8 @@ namespace Hangfire.Dashboard.Management.v2.Support
 		public static List<JobMetadata> Metadata { get; private set; } = new List<JobMetadata>();
 		internal static List<ManagementPageAttribute> Pages { get; set; } = new List<ManagementPageAttribute>();
 
+		internal static ClientSideConfigurations ClientSideConfigurationOptions { get; set; } = new ClientSideConfigurations();
+
 		internal static void GetAllJobs(Assembly assembly)
 		{
 			foreach (Type ti in assembly.GetTypes().Where(x => typeof(IJob).IsAssignableFrom(x) && x.Name != (typeof(IJob).Name)))