+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
-
-## [2.3.2] - 2024-1-11
+## [2.3.2] - 2024-1-17
 
 Minor updates and tweaks.
 
@@ -12,6 +10,13 @@ Minor updates and tweaks.
 - Migrated javascript libraries to use npm instead of ones included in the library project.
 - Altered the way that client resources are loaded using a new dispatcher class
 - Added `connect-src` option to CSP for ASP.Net Core's hot reloading used in the example site.
+
+## Added
+- Added configuration options for client side DateTime controls and "Task Type: Scheduled" execution option
+  - See ASP.Net Core Web Application's \Startup.cs and \HangfireManagement\Expedited.cs files for examples.
+
+## Fixed
+- Scheduling a job when the `DisplayName` attribute is used with non-route friendly characters.
 
 ## [2.3.1] - 2024-1-10